\name{cancer.rats}
\alias{cancer.rats}

\title{Rats data for frailty modeling}

\description{

50 litters of rats were exposed to a carcinogen and observed for time
to tumor occurrence. 

}

\format{

\tabular{ll}{
\code{litter} \tab Number \cr
\code{rx} \tab Exposure indicator (Carcinogen = 1, Control = 0) \cr
\code{time} \tab Tumor occurrence (Days) \cr
\code{event} \tab Indicator if tumor observed (1 = event, 0 =
censored)
}
}

\references{
Mantel, N., Bohidar, N. R., and Ciminera, J. L. (1977), "Mantel-Haenszel analyses
of litter-matched time-to-response data, with modifications for recovery of
interlitter information," \emph{Cancer Research}, 37, 3863-8.
}

\author{S. Kovalchik \email{skoval@ucla.edu}}

