\name{ipdmeta}
\alias{ipdmeta}
\docType{package}

\title{
Mixed Model for Survival Meta-analysis
}
\description{
Functions to perform mixed effects meta-analysis with a survival outcome. Methods for a complete IPD dataset or a partial IPD dataset and aggregate data in the form of survival estimates from a KM plot.
}

\details{
\tabular{ll}{
Package: \tab ipdmeta\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-24\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
 
Maintainer: Stephanie A. Kovalchik <skoval@ucla.edu>

}

\keyword{ package }
