\name{poynard}
\alias{poynard}

\title{Meta-analysis data set for Poynard et al. review of beta-adrenergic-antagonist drugs}

\description{

Meta-analytic dataset of 4 trials of Beta-adrenergic-antagonist drugs in the prevention of gastrointestinal bleeding in
  patients with cirrhosis and esophageal varices.

Dataset used to illustrate power estimator \code{\link{ipd.sep}}
}


\format{

\tabular{ll}{
\code{study} \tab Name of trial \cr
\code{n0} \tab Control group sample size \cr
\code{n1} \tab Beta-adrenergic-antagonist group sample size \cr
\code{bleed0} \tab Control group number of bleeding events at 2 year follow-up \cr
\code{bleed1} \tab Beta-adrenergic-antagonist group number of bleeding events at 2 year follow-up \cr
\code{age0} \tab Control group mean age \cr
\code{age1} \tab Beta-adrenergic-antagonist group mean age \cr
\code{age.s20} \tab Control group variance age \cr
\code{age.s21} \tab Beta-adrenergic-antagonist group variance age \cr
\code{male0} \tab Control group sample proportion male \cr
\code{male1} \tab Beta-adrenergic-antagonist sample proportion male \cr
\code{drug} \tab Beta-adrenergic-antagonist studied 
 }
}

\references{

Poynard, T, Cales, P, Pasta, L, Ideo, G, Pascal, J P, Pagliaro, L, Lebrec, D, (1991),
  Beta-adrenergic-antagonist drugs in the prevention of gastrointestinal bleeding in
  patients with cirrhosis and esophageal varices. An analysis of data and prognostic
  factors in 589 patients from four randomized clinical trials. Franco-Italian
  Multicenter Study Group, \emph{NEJM}, 324, 1532-8.
}


\author{S. Kovalchik \email{skoval@ucla.edu}}

