\name{ipdmeta}
\alias{ipdmeta}
\docType{package}

\title{
Tools for subgroup analyses with multiple trial data using aggregate statistics
}
\description{
This package provides functions to estimate an IPD linear mixed effects model for a continuous outcome and any categorical covariate from study summary statistics. There are also functions for estimating the power of a treatment-covariate interaction test in an individual patient data meta-analysis from aggregate data.
}

\details{
\tabular{ll}{
Package: \tab ipdmeta\cr
Type: \tab Package\cr
Version: \tab 2.3\cr
Date: \tab 2012-11-06\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
}

\author{
 
Maintainer: Stephanie A. Kovalchik <kovalchiksa@nih.gov>

}

\keyword{ package }
