\name{ipdmeta}
\alias{ipdmeta}
\docType{package}

\title{
Functions for characterizing subgroup analyses with meta-analysis
}
\description{
Assess heterogeneity of covariates defining subgroups; estimate the power of a treatment-covariate interaction for an individual patient data meta-analysis using aggregate data
}

\details{
\tabular{ll}{
Package: \tab ipdmeta\cr
Type: \tab Package\cr
Version: \tab 2.2\cr
Date: \tab 2010-1-02\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
}
\author{
 
Maintainer: Stephanie A. Kovalchik <s.a.kovalchik@gmail.com>

}

\keyword{ package }
