\name{ipdmeta}
\alias{ipdmeta}
\docType{package}

\title{
Individual patient and Mixed-level Meta-analysis of
Time-to-Event Outcomes
}
\description{
Functions to estimate the power of an IPD meta-analysis covariate-treatment
interaction (continuous or binary outcome), perform complete IPD meta-analysis and mixed-level
meta-analysis with a survival outcome. 
}

\details{
\tabular{ll}{
Package: \tab ipdmeta\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2010-10-04\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
}
\author{
 
Maintainer: Stephanie A. Kovalchik <skoval@ucla.edu>

}

\keyword{ package }
